feat: LottoTicket 클래스에서 무작위 숫자 만들어내는 로직과의 결합도 낮추기
