+++ src/test/java/lotto/model/lotto/ticket/LottoTicketTest.java
@@ -10,30 +10,20 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 class LottoTicketTest {
-    static class LottoTicketForTest extends LottoTicket {
 
-        Set<LottoNumber> numberMadeSet() {
-            return new HashSet<>(lottoNumbers);
-        }
-
-        List<LottoNumber> lottoNumbers() {
-            return lottoNumbers;
-        }
-    }
-
-    @Nested
-    @DisplayName("LottoTicket 생성자 테스트")
-    class Constructor {
-        @RepeatedTest(value = 100)
-        @DisplayName("한 장의 로또에 있는 숫자는 총 6개이고, 어떤 숫자도 서로 중복이 없다.")
-        void success() {
-            final LottoTicketForTest lottoTicket = new LottoTicketForTest();
-            assertAll(
-                    () -> assertThat(lottoTicket.lottoNumbers().size()).isEqualTo(LottoTicketForTest.NUMBER_COUNT_IN_SINGLE_LOTTO_TICKET),
-                    () -> assertThat(lottoTicket.lottoNumbers().size()).isEqualTo(lottoTicket.numberMadeSet().size())
-            );
-        }
-    }
+//    @Nested
+//    @DisplayName("LottoTicket 생성자 테스트")
+//    class Constructor {
+//        @RepeatedTest(value = 100)
+//        @DisplayName("한 장의 로또에 있는 숫자는 총 6개이고, 어떤 숫자도 서로 중복이 없다.")
+//        void success() {
+//            final LottoTicketForTest lottoTicket = new LottoTicketForTest();
+//            assertAll(
+//                    () -> assertThat(lottoTicket.lottoNumbers().size()).isEqualTo(LottoTicketForTest.NUMBER_COUNT_IN_SINGLE_LOTTO_TICKET),
+//                    () -> assertThat(lottoTicket.lottoNumbers().size()).isEqualTo(lottoTicket.numberMadeSet().size())
+//            );
+//        }
+//    }
 
 //    @Nested
 //    @DisplayName("sameNumbersCount 메서드 테스트")
